Feat - 중간 수정

+++ app/src/main/java/com/demo/sharingapp/domain/user/data/MyInfoData.kt
@@ -4,5 +4,5 @@ data class MyInfoData(
     val profile: String,
     val nickname: String,
     val intro: String,
-    val thumbsUpScore: Int
+    val thumbsUpScore: Double
 )
